refactor: user,business service layer 의존성 리팩토링(#416)
