+++ backend/src/main/java/com/minibeit/businessprofile/web/BusinessProfileController.java
@@ -1,7 +1,7 @@
 package com.minibeit.businessprofile.web;
 
-import com.minibeit.businessprofile.dto.BusinessProfileRequest;
-import com.minibeit.businessprofile.dto.BusinessProfileResponse;
+import com.minibeit.businessprofile.service.dto.BusinessProfileRequest;
+import com.minibeit.businessprofile.service.dto.BusinessProfileResponse;
 import com.minibeit.businessprofile.service.BusinessProfileService;
 import com.minibeit.common.dto.ApiResult;
 import com.minibeit.security.userdetails.CurrentUser;
