+++ backend/src/main/java/com/minibeit/post/domain/repository/PostRepository.java
@@ -4,5 +4,5 @@ import com.minibeit.post.domain.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostRepository extends JpaRepository<Post, Long>, PostRepositoryCustom {
-    boolean existsByBusinessProfileId(Long businessProfileId);
+
 }
