+++ backend/src/main/java/com/minibeit/user/service/UserService.java
@@ -1,29 +1,26 @@
 package com.minibeit.user.service;
 
-import com.minibeit.file.domain.Avatar;
-import com.minibeit.file.service.AvatarService;
-import com.minibeit.businessprofile.domain.repository.BusinessProfileRepository;
 import com.minibeit.common.exception.DuplicateException;
 import com.minibeit.common.exception.InvalidOperationException;
+import com.minibeit.file.domain.Avatar;
+import com.minibeit.file.service.AvatarService;
 import com.minibeit.school.domain.School;
 import com.minibeit.school.domain.SchoolRepository;
 import com.minibeit.user.domain.User;
 import com.minibeit.user.domain.UserVerificationCode;
 import com.minibeit.user.domain.repository.UserRepository;
 import com.minibeit.user.domain.repository.UserVerificationCodeRepository;
+import com.minibeit.user.service.dto.UserRequest;
+import com.minibeit.user.service.dto.UserResponse;
 import com.minibeit.user.service.exception.SchoolNotFoundException;
 import com.minibeit.user.service.exception.UserNotFoundException;
 import com.minibeit.user.service.exception.UserVerificationCodeNotFoundException;
-import com.minibeit.user.service.dto.UserRequest;
-import com.minibeit.user.service.dto.UserResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.stream.Collectors;
-
-;
 
 @Service
 @Transactional
@@ -33,7 +30,6 @@ public class UserService {
     private final SchoolRepository schoolRepository;
     private final UserVerificationCodeRepository userVerificationCodeRepository;
     private final AvatarService avatarService;
-    private final BusinessProfileRepository businessProfileRepository;
 
     public UserResponse.CreateOrUpdate signup(UserRequest.Signup request, User user) {
         nickCheck(request.getNickname());
@@ -96,7 +92,7 @@ public class UserService {
     }
 
     public void deleteOne(User user) {
-        if (!businessProfileRepository.findAllByAdminId(user.getId()).isEmpty()) {
+        if (userRepository.existsBusinessAdminUserById(user.getId())) {
             throw new InvalidOperationException("관리자로 있는 비즈니스 프로필이 존재합니다.");
         }
         avatarService.deleteOne(user.getAvatar());
